refactor physnet output block with residual mlp
